feat(tlb/documentation): add hover documentation for TL-B types
Fixes #422
