+++ modules/asm/src/org/ton/intellij/asm/ide/documentation/AsmDocumentationUtils.kt
@@ -1,0 +1,9 @@
+package org.ton.intellij.asm.ide.documentation
+
+import org.ton.intellij.asm.ide.AsmColor
+import org.ton.intellij.util.doc.DocumentationUtils
+
+object AsmDocumentationUtils : DocumentationUtils() {
+    val asAsmInstruction = loadKey(AsmColor.INSTRUCTION.textAttributesKey)
+    val asComment = loadKey(AsmColor.COMMENT.textAttributesKey)
+}
