+++ modules/tolk/src/org/ton/intellij/tolk/doc/TolkDocumentationProvider.kt
@@ -7,27 +7,27 @@ import com.intellij.psi.PsiDocCommentBase
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.SyntaxTraverser
-import org.ton.intellij.tolk.doc.DocumentationUtils.appendNotNull
-import org.ton.intellij.tolk.doc.DocumentationUtils.asAnnotation
-import org.ton.intellij.tolk.doc.DocumentationUtils.asComma
-import org.ton.intellij.tolk.doc.DocumentationUtils.asConstant
-import org.ton.intellij.tolk.doc.DocumentationUtils.asDot
-import org.ton.intellij.tolk.doc.DocumentationUtils.asField
-import org.ton.intellij.tolk.doc.DocumentationUtils.asFunction
-import org.ton.intellij.tolk.doc.DocumentationUtils.asGlobalVariable
-import org.ton.intellij.tolk.doc.DocumentationUtils.asIdentifier
-import org.ton.intellij.tolk.doc.DocumentationUtils.asKeyword
-import org.ton.intellij.tolk.doc.DocumentationUtils.asNumber
-import org.ton.intellij.tolk.doc.DocumentationUtils.asParameter
-import org.ton.intellij.tolk.doc.DocumentationUtils.asParen
-import org.ton.intellij.tolk.doc.DocumentationUtils.asPrimitive
-import org.ton.intellij.tolk.doc.DocumentationUtils.asString
-import org.ton.intellij.tolk.doc.DocumentationUtils.asStruct
-import org.ton.intellij.tolk.doc.DocumentationUtils.asTypeAlias
-import org.ton.intellij.tolk.doc.DocumentationUtils.asTypeParameter
-import org.ton.intellij.tolk.doc.DocumentationUtils.colorize
-import org.ton.intellij.tolk.doc.DocumentationUtils.line
-import org.ton.intellij.tolk.doc.DocumentationUtils.part
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.appendNotNull
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asAnnotation
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asComma
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asConstant
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asDot
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asField
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asFunction
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asGlobalVariable
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asIdentifier
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asKeyword
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asNumber
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asParameter
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asParen
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asPrimitive
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asString
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asStruct
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asTypeAlias
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.asTypeParameter
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.colorize
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.line
+import org.ton.intellij.tolk.doc.TolkDocumentationUtils.part
 import org.ton.intellij.tolk.doc.psi.TolkDocComment
 import org.ton.intellij.tolk.highlighting.TolkSyntaxHighlighter
 import org.ton.intellij.tolk.psi.TOLK_KEYWORDS
@@ -47,7 +47,6 @@ import org.ton.intellij.tolk.psi.TolkSelfTypeExpression
 import org.ton.intellij.tolk.psi.TolkStruct
 import org.ton.intellij.tolk.psi.TolkStructField
 import org.ton.intellij.tolk.psi.TolkTypeDef
-import org.ton.intellij.tolk.psi.TolkTypeExpression
 import org.ton.intellij.tolk.psi.TolkTypeParameter
 import org.ton.intellij.tolk.psi.TolkTypeParameterList
 import org.ton.intellij.tolk.psi.TolkVar
