+++ modules/util/src/org/ton/intellij/util/doc/DocumentationUtils.kt
@@ -1,4 +1,4 @@
-package org.ton.intellij.asm.ide.documentation
+package org.ton.intellij.util.doc
 
 import com.intellij.lang.documentation.DocumentationSettings
 import com.intellij.openapi.editor.colors.EditorColorsManager
@@ -6,14 +6,10 @@ import com.intellij.openapi.editor.colors.TextAttributesKey
 import com.intellij.openapi.editor.markup.TextAttributes
 import com.intellij.openapi.editor.richcopy.HtmlSyntaxInfoUtil
 import io.ktor.util.*
-import org.ton.intellij.asm.ide.AsmColor
 
-object DocumentationUtils {
-    private fun loadKey(key: TextAttributesKey): TextAttributes =
+open class DocumentationUtils {
+    protected fun loadKey(key: TextAttributesKey): TextAttributes =
         EditorColorsManager.getInstance().globalScheme.getAttributes(key)!!
-
-    val asAsmInstruction = loadKey(AsmColor.INSTRUCTION.textAttributesKey)
-    val asComment = loadKey(AsmColor.COMMENT.textAttributesKey)
 
     @Suppress("UnstableApiUsage")
     fun StringBuilder.colorize(text: String, attrs: TextAttributes, noHtml: Boolean = false) {
@@ -28,12 +24,45 @@ object DocumentationUtils {
         )
     }
 
+    fun StringBuilder.part(text: String?) {
+        if (text.isNullOrEmpty()) {
+            return
+        }
+        append(text)
+        append(" ")
+    }
+
+    fun StringBuilder.part(text: String?, attrs: TextAttributes) {
+        if (text == null) {
+            return
+        }
+        colorize(text, attrs)
+        append(" ")
+    }
+
+    fun StringBuilder.appendNotNull(text: String?) {
+        if (text == null) {
+            return
+        }
+        append(text)
+    }
+
     fun StringBuilder.line(text: String?) {
         if (text == null) {
             return
         }
         append(text)
         append("\n")
+    }
+
+    fun StringBuilder.monospaced(text: String, attrs: TextAttributes? = null, noHtml: Boolean = false) {
+        append("<code>")
+        if (attrs != null) {
+            colorize(text, attrs, noHtml)
+        } else {
+            append(text)
+        }
+        append("</code>")
     }
 
     fun colorize(text: String, attrs: TextAttributes, noHtml: Boolean = false): String {
